fix: change primitive to boxed analogs on the builder's methods on body-classes.
fix: Timetracking fields change to double type.

+++ lib/src/main/java/io/github/patbattb/yougileapilib/domain/TimeTracking.java
@@ -19,9 +19,9 @@ public class TimeTracking {
     /**
      * How many hours were scheduled to complete the task.
      */
-    int plan;
+    double plan;
     /**
      * How many hours were spent on the task.
      */
-    int work;
+    double work;
 }
